FILES, FOLDERS AND COMPONENTS ADDED
Parcel version updated to enable functionality

folder and file structure for movie-view, movie-card and main-view created

code for components added to enable the movie names to be displayed and for additional information about the movies

code added so the main view can import the movie-card component and the information can be displayed in a movie view.

workable links to images created, descriptions updated and directors added.

Site tested and ran on localhost1234

README file updated